refactor(extractor): Move extractors to entity-registry (#4307)
* Move extractors to entity-registry

* Fix test

* Fix checkstyle
